from 19; started /me/settings/uploadAvatar
