+++ srcs/database/package.json
@@ -16,6 +16,7 @@
   "license": "ISC",
   "dependencies": {
     "@fastify/autoload": "^6.0.0",
+    "@fastify/multipart": "^9.0.3",
     "@fastify/sensible": "^6.0.0",
     "axios": "^1.9.0",
     "bcrypt": "^5.1.1",
