+++ srcs/database/routes/avatars/schemas/loader.js
@@ -4,4 +4,5 @@ const fp = require('fastify-plugin')
 
 module.exports = fp(async function schemaLoaderPlugin (fastify, opts) {
   fastify.addSchema(require('./getAvatar.json'))
+  fastify.addSchema(require('./avatar.json'))
 })
