+++ srcs/database/routes/me/autohooks.js
@@ -3,7 +3,6 @@
 
 const fp = require('fastify-plugin')
 const schemas = require('./schemas/loader')
-const fileTypeFromBuffer = require('file-type')
 
 
 module.exports = fp(async function meAutoHooks (fastify, opts) {
@@ -45,30 +44,10 @@ module.exports = fp(async function meAutoHooks (fastify, opts) {
         fastify.log.error(`getUserProfile error: ${err.message}`)
         throw new Error('User profile retrieval failed')
       }
-    },
+    }
     
-    async createAvatar(userId, avatar) {
-      try {
-        const avatarBuffer = Buffer.isBuffer(avatar) ? avatar : Buffer.from(avatar)
-        const type = await fileTypeFromBuffer(avatarBuffer)
-        if (!type || !['image/jpeg', 'image/png', 'image/jpg'].includes(type.mime)) {
-          throw new Error(`Unsupported image format: ${type?.mime || 'unknown'}`)
-        }    
-        const mimeType = type.mime
-        const query = fastify.db.prepare(`
-          INSERT INTO user_avatars (user_id, avatar, mime_type)
-          VALUES (?, ?, ?)
-        `)
-        const result = query.run(userId, avatarBuffer, mimeType)
-        fastify.log.debug(`createAvatar: ${userId} -> ID ${result.lastInsertRowid}`)
-        return result.lastInsertRowid
-      } catch (err) {
-        fastify.log.error(`createAvatar error: ${err.message}`)
-        throw new Error('Avatar creation failed')
-      }
-    },
   })
 }, {
   name: 'meAutoHooks',
-  dependencies: ['database', 'defaultAssets']
+  dependencies: ['database']
 })
