+++ srcs/database/routes/me/schemas/loader.js
@@ -3,7 +3,6 @@
 const fp = require('fastify-plugin')
 
 module.exports = fp(async function schemaLoaderPlugin (fastify, opts) {
-  fastify.addSchema(require('./avatar.json'))
   fastify.addSchema(require('./getProfile.json'))
   fastify.addSchema(require('./userProfile.json'))
 })
