+++ srcs/user_management/routes/user/schemas/avatar.json
@@ -14,8 +14,7 @@
             "type":  { "const": "Buffer" },
             "data":  { 
               "type": "array", 
-              "items": { "type": "integer" 
-              } 
+              "items": { "type": "integer" } 
             }
         }}
       ]
