+++ srcs/user_management/routes/user/schemas/loader.js
@@ -4,4 +4,5 @@ const fp = require('fastify-plugin')
 
 module.exports = fp(async function schemaLoaderPlugin (fastify, opts) {
   fastify.addSchema(require('./userProfile.json'))
+  fastify.addSchema(require('./avatar.json'))
 })
